requirements.txt
click==6.6
coreapi==1.32.3
coreapi-cli==1.0.0
Django==1.10
djangorestframework==3.4.5
itypes==1.1.0
Pygments==2.1.3
requests==2.11.1
simplejson==3.8.2
uritemplate==0.6
